Buffer request body before sending to http.request.
Looking at the Channels code, it (understandably) blocks and reads all
body chunks before continuing to process the request. We want to avoid
the Django processes blocking while waiting for data, so it's best for
us to read the entire body into memory before starting to send it down
the http.request channel.

This has the nice side-effect that we can simplify how we send the
chunks: we no longer need to have an empty chunk at the end to signal
more_content=false. As we have the complete vec in memory, we can simply
chunk it and then peek ahead in the iterator. We can simplify the chain
of futures by just doing it all in send_request().

FWIW, Daphne doesn't even chunk the body of requests right now.