Make deserializing a response a two-step process.
When waiting on more than one channel a caller might not know what reply
they expect (or rather they may expect the reply to be one of a known
number of types). Return the buffer to the user (wrapped in a type so
they ought to know not to mess with it) and then provide a separate
method for deserialization. (The hope is that once the user knows what
channel it came from they will know what type of message they've got).

This seems like it's a bit of a pain for callers, but it will be hidden
away once we have a separate thread to pump replies for a collection of
channels.