Listen for replies in a separate thread.
Have a separate thread that periodically polls the channel layer for
replies on a set of channels that we're subscribed to. Allow tasks
(requests) to request that we subscribe to a channel and wake it up when
we receive a message from that channel.

It would be nice if we didn't have to poll redis, but it's not clear how
to do a blocking receive whilst also allowing new subscriptions. At
least do better than asgi_redis by avoiding the busy-loop to check if we
have new subscriptions, by blocking on the queue that other tasks use to
send us subscription requests.

Phew, that was more complex than I thought it was going to be!