Tidy up error handling in main Service.
Don't bother trying to pass a hyper::Error through - it's meaningless
for our application. For now we want to just propogate the precense of
an error, then turn this into an error response (which is a successfuly
error as far as Hyper is concerned) before returning.

For now we will use () as our error-type, but we may wish to have an
error enum in future, if we want special handling for the "channel full"
error.